% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{extract_coefs}
\alias{extract_coefs}
\title{Extract and Format Model Coefficients}
\usage{
extract_coefs(x, digit = 3)
}
\arguments{
\item{x}{A model object generated from \code{catalytic} that containing model coefficients.}

\item{digit}{An integer specifying the number of decimal places for rounding coefficients. Default is 3.}
}
\value{
A named numeric vector of model coefficients, rounded to the specified number of decimal places.
}
\description{
This function retrieves the coefficients from a \code{x} object, formats them
with appropriate names, and rounds each coefficient to the specified number of decimal places.
Optionally, the intercept can be included or excluded from the output.
}
