% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{extract_dim}
\alias{extract_dim}
\title{Extract Dimension Information from Model Initialization}
\usage{
extract_dim(cat_init)
}
\arguments{
\item{cat_init}{A list containing model initialization data, expected to include
\code{obs_size} (observed data size), \code{syn_size} (synthetic data size), \code{size} (total data size),
and \code{x} (the covariate matrix).}
}
\value{
A character string summarizing the dimensions of the dataset used in the model.
}
\description{
This function retrieves and formats the dimensions of the dataset used in the model,
including the number of observed and synthetic data points and the total number of rows
and columns.
}
