% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_adjusted_cat_init}
\alias{get_adjusted_cat_init}
\title{Adjusted Cat Initialization}
\usage{
get_adjusted_cat_init(cat_init, formula_rhs)
}
\arguments{
\item{cat_init}{The object generated from \code{cat_glm_initialization}, \code{cat_cox_initialization} or \code{cat_lmm_initialization}}

\item{formula_rhs}{A formula specifying the right-hand side of the model for predictors.}
}
\value{
A list containing the original \code{cat_init} with added components:
- \code{adj_x}: The adjusted model frame for the predictors.
- \code{adj_obs_x}: The observed part of the adjusted predictors.
- \code{adj_syn_x}: The synthetic part of the adjusted predictors.
}
\description{
This function adjusts the categorical initialization by creating a model frame
for the predictors specified in the right-hand side of the formula and splits
the adjusted data into observed and synthetic parts.
}
