% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_cox_kappa}
\alias{get_cox_kappa}
\title{Estimate the kappa value for the synthetic Cox proportional hazards model}
\usage{
get_cox_kappa(X, time, status, hazard_constant)
}
\arguments{
\item{X}{A matrix of covariates with rows representing observations and columns representing features.}

\item{time}{A vector of time-to-event data.}

\item{status}{A vector indicating event occurrence (1 = event, 0 = censored).}

\item{hazard_constant}{A scalar representing the hazard constant. Defaults to NULL, in which case it's calculated internally.}
}
\value{
A numeric value representing the estimated kappa for the synthetic Cox model.
}
\description{
This function iterative estimates the kappa value for the synthetic Cox proportional hazards model using a vectorized approach for efficiency.
}
