% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_formula_string}
\alias{get_formula_string}
\title{Convert Formula to String}
\usage{
get_formula_string(formula)
}
\arguments{
\item{formula}{A formula object to be converted to a string.}
}
\value{
A character string representing the formula.
}
\description{
This function converts a formula object to a character string. It removes extra whitespace and formats the formula as a single line.
}
