% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_glm_diag_approx_cov}
\alias{get_glm_diag_approx_cov}
\title{Compute Diagonal Approximate Covariance Matrix}
\usage{
get_glm_diag_approx_cov(X, model)
}
\arguments{
\item{X}{Matrix. The design matrix (predictors) for the GLM.}

\item{model}{A fitted GLM model object. The object should contain the fitted values
and prior weights necessary for computing the Hessian.}
}
\value{
Numeric vector. The diagonal elements of the approximate covariance matrix.
}
\description{
This function computes the diagonal elements of the approximate covariance matrix
for the coefficients in a generalized linear model (GLM). The covariance is derived
from the second derivative (Hessian) of the log-likelihood function.
}
