% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_glm_family_string}
\alias{get_glm_family_string}
\title{Retrieve GLM Family Name or Name with Link Function}
\usage{
get_glm_family_string(family, with_link = FALSE)
}
\arguments{
\item{family}{Character or function. The name of the GLM family (as a string) or a function that returns a GLM family object.}

\item{with_link}{Logical. If TRUE, returns the family name along with the link function in the format "family \[link\]".
If FALSE, only the family name is returned. Default is FALSE.}
}
\value{
A character string. The name of the GLM family, or the name with the link function if \code{with_link} is TRUE.
}
\description{
This function retrieves the name of a GLM family or, optionally, the family name with the associated link function.
}
