% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_glm_sample_data}
\alias{get_glm_sample_data}
\title{Generate Sample Data for GLM}
\usage{
get_glm_sample_data(family_string, n = 10, mean = 0, sd = 1)
}
\arguments{
\item{family_string}{Character. The family of the GLM. Options are \code{"binomial"} or \code{"gaussian"}.}

\item{n}{Integer. The number of samples to generate.}

\item{mean}{Numeric. The mean of the distribution (used for both binomial and Gaussian).}

\item{sd}{Numeric. The standard deviation of the distribution (used only for Gaussian).}
}
\value{
Numeric vector of generated sample data.
}
\description{
This function generates sample data for a specified GLM family. It can generate
binomial or Gaussian distributed data based on the provided parameters.
}
