% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_linear_predictor}
\alias{get_linear_predictor}
\title{Compute Linear Predictor}
\usage{
get_linear_predictor(X, coefs)
}
\arguments{
\item{X}{A matrix of predictor variables.}

\item{coefs}{A vector of coefficients. It should be either the same length as the number of columns in X
(for models without an intercept) or one more than the number of columns in X (for models with an intercept).}
}
\value{
A vector of linear predictor values.
}
\description{
This function computes the linear predictor from a matrix of predictor variables and a vector of coefficients.
It handles cases with and without an intercept term.
}
