% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print_df_head_tail}
\alias{print_df_head_tail}
\title{Print Data Frame with Head and Tail Rows}
\usage{
print_df_head_tail(df, digit = 3)
}
\arguments{
\item{df}{A data frame to display.}

\item{digit}{An integer specifying the number of decimal places to which the summary
statistics should be rounded. Default is 3.}
}
\value{
Invisibly returns the original data frame.
}
\description{
This function displays the first 5 and last 5 rows of a data frame.
Column names are displayed only for the first 5 rows, with ellipses (\code{...})
in the middle to indicate additional rows.
}
